requirements.txt
Django>=3.0.8,<4.0.0
djangorestframework>=3.11.0,<4.0.0
channels>=2.4.0,<3.0.0
Pillow>=7.2.0,<8.0.0
psycopg2>=2.8.5,<2.9.0
pytest-asyncio>=0.14.0,<1.0.0
pytest-django>=3.9.0,<4.0.0
django-trench>=0.2.3,<1.0.0
uwsgi>=2.0.18,<2.1.0
